backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymysql==1.1.0
snowflake-connector-python==3.6.0
pyarrow==15.0.0
sas7bdat==2.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.3

# AI/ML
ollama==0.1.6
google-generativeai==0.8.6
google-auth==2.27.0
duckdb==0.9.2
chromadb==0.4.22
sqlglot==20.8.0
polars==0.20.5
sentence-transformers==2.3.1
torch==2.1.2
vl-convert-python==1.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-cov==4.1.0
httpx>=0.25.2,<0.26.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.1
